requirements.txt
Flask>=2.0.0
aiohttp==3.9.3
beautifulsoup4==4.12.2
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2023.3